fix(ssh): Fix ssh-agent forwarding
If the current machine is a desktop, tell ssh-agent to piggyback on the
gpg-agent for authentication. This means setting SSH_AUTH_SOCK to
S.gpg-agent.ssh. Whereas, during ssh-agent forwarding, the server
machine should use the default socket provided by the sshd daemon.

Hence, the condition to check if the current system is a desktop or a
server has been added. Now the SSH_AUTH_SOCK is set to use the gpg-agent
on S.gpg-agent.ssh only when there is a keygrip specified in the
~/.gnupg/sshcontrol file.